Increase client timeout and adjust chunk size
Increased the client timeout from default 2 minutes to 4 minutes in the TTSApi Program.cs - this is to accomodate longer texts that need more processing time on Salute Speech Service. Reduced the number of sentences per chunk from 30 to 20 in the Reader Index.razor - this modification ensures better management of larger texts and a smoother reading experience.