Update Dockerfile and adjust certificate path in settings
Updated Dockerfile to use a multi-stage build process which optimizes the image size and makes sure to copy only the necessary files. Also adjusted the certificate path in the appsettings.json file to match the new directory structure in the docker image. Significantly, the installation of ffmpeg and the apt-get update is combined into a single line to further reduce the number of layers in the Docker image.