Add logging to TextToSpeech post endpoint
This commit introduces several logging statements to the TextToSpeech post endpoint in order to facilitate debugging. Specifically, it logs when a new request is received, when a request with empty text is received, after the text has been sanitized, when the audio has been generated successfully and when the generated audio file has been opened successfully. It also includes a small refactor renaming error variable 'e' to 'exception' for more semantic correctness.