Add logging to DeleteOnCloseStream class
Incorporated ILogger to DeleteOnCloseStream class. Now, the logger will track when a new instance of DeleteOnCloseStream is initialized, delete a file upon dispose, and log any errors that occur during the file deletion process. This change was made to provide greater transparency into the operations of the DeleteOnCloseStream instances, which will make debugging and issue tracking easier in the future.