Add click-outside functionality to close modal
Added a functionality to the modal window so it can be closed when the user clicks outside of the window. Modified the modal div with a click event handler, and then introduced a new method 'HandleModalClick' to handle the closing action. Also, to prevent the propagation of the click event to the 'modal-window' div, an 'onclick:stopPropagation' attribute was added.